Refactor, document migration utilities
Slightly rework, and document, and rename the migration scripts and
supporting utility, for clarity.
